V-02 Order page adaptive
