+++ src/components/buttons/buttonPrimary/styles.scss
@@ -2,7 +2,7 @@
 @import '../../../scss/variables';
 
 .button-primary {
-  width: 250px;
+  width: 100%;
   height: 50px;
   border-radius: 8px;
   transition: $transition-sm;
