+++ src/components/forms/locationForm/index.tsx
@@ -5,11 +5,12 @@ import InputPrimary from '../../inputPrimary'
 
 export default function SearchLocationForm() {
   return (
-    <form
-      className='search-location-form'
-    >
+    <form className='search-location-form'>
       <div className='search-location-form__form-wrapper'>
-        <label htmlFor="townPicker">Город</label>
+        <label htmlFor="townPicker">
+          Город
+        </label>
+
         <InputPrimary
           name="townPicker"
           placeholder='Начните вводить город...'
@@ -17,7 +18,10 @@ export default function SearchLocationForm() {
       </div>
 
       <div className='search-location-form__form-wrapper'>
-        <label htmlFor="pickUpPoint">Пункт выдачи</label>
+        <label htmlFor="pickUpPoint">
+          Пункт выдачи
+        </label>
+
         <InputPrimary
           name="pickUpPoint"
           placeholder='Начните вводить пункт...'
