+++ src/components/forms/locationForm/styles.scss
@@ -10,10 +10,30 @@
     justify-content: flex-end;
 
     label {
+      display: inline-block;
       margin-right: $space-sm;
+      overflow: hidden;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+      max-width: 95px;
       font-family: 'RobotoLight';
       @extend %heading14;
       color: $primary-color;
+    }
+  }
+
+  @media (max-width: getBreackpoint('sm')) {
+    &__form-wrapper {
+      label {
+        font-size: 13px;
+        line-height: 15px;
+      }
+    }
+  }
+
+  @media (max-width: 319px) {
+    &__form-wrapper {
+      width: 93%;
     }
   }
 }
