+++ src/components/forms/orderField/styles.scss
@@ -8,7 +8,7 @@
 }
 
 .order-field {
-  width: 80%;
+  width: 100%;
   position: relative;
   margin: .5rem 0;
   background-color: $white-color;
