+++ src/components/header/styles.scss
@@ -1,9 +1,11 @@
 @import '../../scss/typography';
 @import '../../scss/variables';
+@import '../../scss/mixins';
 
 .header {
-  display: flex;
-  align-items: center;
+  width: 100%;
+  height: 100%;
+  @include flex-align-center;
   justify-content: space-between;
 
   &__title {
@@ -12,6 +14,7 @@
   }
 
   @media (max-width: getBreackpoint('sm')) {
+    height: auto;
     flex-direction: column;
     align-items: flex-end;
     padding: $space-md;
