+++ src/routes/orderPage/styles.scss
@@ -6,17 +6,14 @@
   height: 100vh;
   display: grid;
   grid-template-columns: $aside-width 1fr;
-  overflow: hidden;
-
-  &__header {
-    margin: 0 15px;
-  }
+  overflow-y: auto;
+  overflow-x: hidden;
 
   &__main {
     display: grid;
     width: 100%;
     height: 100%;
-    grid-template-rows: 100px 32px 1fr;
+    grid-template-rows: 100px auto 1fr;
   }
 
   &__breadcrumbs {
@@ -31,14 +28,17 @@
     display: flex;
 
     &--left {
-      flex: 2;
-      padding: $space-lg;
+      flex: 2.5;
+      padding-top: $space-lg;
+      padding-right: $space-lg;
       border-right: 1px solid $grey-color;
     }
 
     &--right {
       flex: 1;
       padding: $space-lg;
+      padding-right: 0;
+      padding-bottom: 0;
     }
 
     &--title {
@@ -46,12 +46,13 @@
       font-size: 14px;
       line-height: 16px;
       color: $primary-color;
+      margin-top: $space-xl;
+      margin-bottom: $space-md;
     }
 
     &--map {
       max-width: 735px;
       height: 350px;
-      padding: $space-xl 0;
     }
   }
 
@@ -61,19 +62,71 @@
   }
 
   @media (max-width: getBreackpoint('lg')) {
+    &__main {
+      grid-template-rows: 80px auto 1fr;
+    }
+
     &__content {
       flex-direction: column;
       flex-wrap: wrap;
 
       &--left {
         flex: 1;
-        padding: $space-md;
-        border-right: 1px solid $grey-color;
+        padding: $space-md 0;
+        border-right: none;
       }
   
       &--right {
         flex: 1;
-        padding: $space-md;
+        padding: $space-lg 0;
+      }
+
+      &--map {
+        max-width: 100%;
+        height: 340px;
+      }
+    }
+  }
+
+  @media (max-width: getBreackpoint('md')) {
+    &__content {
+      &--map {
+        height: 300px;
+      }
+    }
+  }
+
+  @media (max-width: getBreackpoint('sm')) {
+    grid-template-columns: 1fr;
+
+    &__main {
+      grid-template-rows: 100px auto 1fr;
+    }
+
+    &__aside {
+      display: none;
+    }
+
+    &__content {
+      &--left {
+        padding: $space-md 0;
+      }
+
+      &--map {
+        height: 280px;
+      }
+    }
+  }
+
+  @media (max-width: getBreackpoint('vsm')) {
+    &__content {
+      &--title {
+        margin-top: $space-lg;
+        margin-bottom: $space-sm;
+      }
+
+      &--map {
+        height: 250px;
       }
     }
   }
