+++ src/scss/_variables.scss
@@ -33,6 +33,7 @@ $aside-width: 64px;
 // Breakpoints
 $grid-breakpoints: (
   tiny: 320px,
+  vsm: 460px,
   sm: 576px,
   md: 768px,
   lg: 1024px,
